+++ package/CHANGELOG.md
@@ -51,3 +51,21 @@
 ### Patch Changes
 
 - [#8](https://github.com/louisescher/typed-rest-routes/pull/8) [`1312b81`](https://github.com/louisescher/typed-rest-routes/commit/1312b81f564653390b1f1394ce32a6de5f51aae0) Thanks [@louisescher](https://github.com/louisescher)! - Fixes types for query parameters and allows for passing custom zod errors
+
+## 0.1.2
+
+### Patch Changes
+
+- [`9a93136`](https://github.com/louisescher/typed-rest-routes/commit/9a93136c98739210f52eb960566cf7296c5782cd) Thanks [@louisescher](https://github.com/louisescher)! - Fixes route types not being autocompleted
+
+## 0.1.1
+
+### Patch Changes
+
+- [#5](https://github.com/louisescher/typed-rest-routes/pull/5) [`10b67c8`](https://github.com/louisescher/typed-rest-routes/commit/10b67c8e4a8cd0de8e717e35e05c45ab173da5fc) Thanks [@louisescher](https://github.com/louisescher)! - Fixes client-side types not working as intended on Linux
+
+## 0.1.0
+
+### Minor Changes
+
+- [#2](https://github.com/louisescher/typed-rest-routes/pull/2) [`c6f74f7`](https://github.com/louisescher/typed-rest-routes/commit/c6f74f7ba77dc1b0089294b44294a5673ab14aaf) Thanks [@louisescher](https://github.com/louisescher)! - Initial Release